+++ recipes/esm2_native_te_mfsdp/Dockerfile
