+++ recipes/esm2_native_te_mfsdp/README.md
@@ -1,0 +1,15 @@
+# ESM-2 training with megatron-fsdp and custom pytorch training loop
+
+Vuild the docker image with the following command:
+
+```bash
+docker build -t my_image .
+```
+
+## Running training
+
+Run training with
+
+```bash
+docker run --rm -it --gpus all my_image torchrun train_mfsdp.py --config-name L0_sanity
+```
