+++ recipes/esm2_native_te_mfsdp/hydra_config/L1_15B_perf_test.yaml
