+++ recipes/esm2_native_te_mfsdp/hydra_config/L1_650M.yaml
