+++ recipes/esm2_native_te_mfsdp/slurm.sh
