+++ recipes/esm2_native_te_mfsdp/train_fsdp2.py
@@ -81,7 +81,7 @@ def get_linear_schedule_with_warmup(
 
 @hydra.main(config_path="hydra_config", config_name="L0_sanity.yaml", version_base="1.2")
 def main(args: DictConfig):
-    """Train ESM-2 with TE layers using nvFSDP.
+    """Train ESM-2 with TE layers using megatron-fsdp.
 
     Model names are valid ESM-2 model sizes, e.g.:
     - "esm2_t6_8M_UR50D"
@@ -90,7 +90,7 @@ def main(args: DictConfig):
     """
     # Initialize distributed training and create a device mesh for FSDP.
     # We have to create a dummy mesh dimension for context parallel and tensor parallel for things
-    # to work correctly with nvFSDP.
+    # to work correctly with megatron-fsdp.
     dist.init_process_group(backend="nccl")
     dist_config = DistributedConfig()
     torch.cuda.set_device(dist_config.local_rank)
